requirements.txt
speedtest_cli==2.1.3
streamlit==1.27.2
